Update README.md
this  is my first major project in a college .And It is a beautiful weather  app in java script . It  has search suggestions when a user enters any city/country. And I add the autocomplete search suggestion in a search bar. It showcases background images based on location data and it displays background images based on particular location. I used two APIs in project open weather API for fetching weather data and for background images used unsplash  API.